HALF FINISHED CHANGES, adding methods to loggers and small edits to progresses to allow checking of goals, integrating goals into rest of system

+++ SoftEndImp_HKF/src/java/Model/PhysicalHealth/Weight.java
@@ -8,7 +8,7 @@ public class Weight {
         this.grams = grams;
     }
 
-    public double getGrams() {
+    public int getGrams() {
         return grams;
     }
